// types/type.ts
export type Attributes = {
  width: string;
  height: string;
  fontSize: string;
  fontFamily: string;
  fontWeight: string;
  fill: string;
  stroke: string;
  strokeWidth: number;
  opacity: number;
};

export type CanvasAttributes = {
  backgroundColor: string;
}

export type ActiveElement = {
  name: string;
  value: string;
  icon: string;
} | null;